Playing around with middleware and channels
