+++ src/Kwtc.ErrorMonitoring.Client.AspNetCore/Kwtc.ErrorMonitoring.Client.AspNetCore.csproj
@@ -15,13 +15,14 @@
         <Description>ASP.NET Core utilities for using .NET client library for Kwtc.ErrorMonitoring.</Description>
         <PackageTags>Error Exception Reporting Kwtc.ErrorMonitoring ASP.NET</PackageTags>
     </PropertyGroup>
-    
+
     <ItemGroup>
         <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
         <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="7.0.0" />
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\Kwtc.ErrorMonitoring.Client\Kwtc.ErrorMonitoring.Client.csproj" />
+        <ProjectReference Include="..\Kwtc.ErrorMonitoring.Client\Kwtc.ErrorMonitoring.Client.csproj" />
     </ItemGroup>
+
 </Project>
