+++ src/Kwtc.ErrorMonitoring.Client.AspNetCore/ServiceCollectionExtensions.cs
@@ -1,3 +1,4 @@
+using System.Threading.Channels;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Kwtc.ErrorMonitoring.Client.AspNetCore;
@@ -6,7 +7,9 @@ public static class ServiceCollectionExtensions
 {
     public static IServiceCollection AddErrorMonitoring(this IServiceCollection services)
     {
-        services.AddScoped<IApiClient, ApiClient>();
+        services
+            .AddScoped<IApiClient, ApiClient>()
+            .AddSingleton(Channel.CreateBounded<ExceptionEvent>(10000));
 
         return services;
     }
